Create component for logo small
